backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pinecone-client==2.2.1
langchain==0.0.184
python-dotenv==0.19.2
PyPDF2==3.0.1
openai==1.0.0